Split data askgproxiidemod to own file
added lf gproxii read (from antenna)
fixed a few copy paste issues in the comments of some lf files
