+++ client/cmddata.h
@@ -23,7 +23,6 @@ int CmdData(const char *Cmd);
 void printDemodBuff(void);
 void setDemodBuf(uint8_t *buff, size_t size, size_t startIdx);
 int CmdPrintDemodBuff(const char *Cmd);
-int CmdG_Prox_II_Demod(const char *Cmd);
 int Cmdaskrawdemod(const char *Cmd);
 int Cmdaskmandemod(const char *Cmd);
 int AutoCorrelate(int window, bool SaveGrph, bool verbose);
