+++ client/cmdlfpresco.h
@@ -4,7 +4,7 @@
 // at your option, any later version. See the LICENSE.txt file for the text of
 // the license.
 //-----------------------------------------------------------------------------
-// Low frequency T55xx commands
+// Low frequency Presco tag commands
 //-----------------------------------------------------------------------------
 #ifndef CMDLFPRESCO_H__
 #define CMDLFPRESCO_H__
